Conflicts:
	vs2010/TestProject/source/TestScene.cpp

+++ src/Audio/Sound.cpp
@@ -1,4 +1,4 @@
-#include <UtH/Audio/Sound.h>
+#include <UtH/Audio/Sound.hpp>
 #include <UtH/Platform/Typedefs.hpp>
 #include <UtH/Platform/Debug.hpp>
 #include <UtH/Resources/ResourceManager.h>
@@ -171,7 +171,7 @@ void Sound::Initialize(const char* fileName)
 		{
 			alBufferData(buffer, AL_FORMAT_STEREO8, 
 				buf.GetSoundInfo().soundBuffer, 
-				buf.GetSoundInfo().frames * sizeof(int), 
+				buf.GetSoundInfo().frames * sizeof(short), 
 				buf.GetSoundInfo().sampleRate);
 		}
 	}
@@ -181,14 +181,14 @@ void Sound::Initialize(const char* fileName)
 		{
 			alBufferData(buffer, AL_FORMAT_MONO16, 
 				buf.GetSoundInfo().soundBuffer, 
-				buf.GetSoundInfo().frames * sizeof(int), 
+				buf.GetSoundInfo().frames * sizeof(short), 
 				buf.GetSoundInfo().sampleRate);
 		}
 		else if(buf.GetSoundInfo().bitsPerSample == 8)
 		{
 			alBufferData(buffer, AL_FORMAT_MONO8, 
 				buf.GetSoundInfo().soundBuffer, 
-				buf.GetSoundInfo().frames * sizeof(int), 
+				buf.GetSoundInfo().frames * sizeof(short), 
 				buf.GetSoundInfo().sampleRate);
 		}
 	}
